Get rid of some warnings and add another 64-bit boost installer
Former-commit-id: 01260eb1eed2b8804b5bc63bdf808d64f6c673c8

+++ Code/Source/Algorithms/KDTree3.hpp
@@ -446,7 +446,7 @@ class /* THEA_API */ KDTree3
         if (filters.empty())
         {
           std::copy(begin, end, elems.begin());
-          num_elems = max_new_elems;
+          num_elems = (long)max_new_elems;
         }
         else
         {
